requirements.txt
google-adk
curl_cffi